pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "living-codex"
version = "0.1.0"
description = "AI-powered GM assistant for TTRPG campaigns"
requires-python = ">=3.10"
dependencies = [
    "discord.py>=2.3.0",
    "aiosqlite>=0.20.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "rapidfuzz>=3.0.0",
    "httpx>=0.27.0",
    "google-generativeai>=0.5.0",
    "google-genai>=1.0.0",
    "anthropic>=0.40.0",
    "python-dotenv>=1.0.0",
    "watchfiles>=0.20.0",
    "markdown>=3.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/living_codex"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I"]
